:recycle: Refactor: 코드리뷰 수정사항 반영(함수 간략화, 폴더구조 변경)
- onClick 이벤트 코드 작성을 간략화
- scrap과 category state, hook 함수를 서로 다른 폴더로 구분
- 사용하지 않는 selector 삭제
