+++ src/application/store/category/categoryHook.ts
@@ -1,0 +1,6 @@
+import { useRecoilValue, useResetRecoilState } from 'recoil';
+
+import { categoryIdsArray } from './categoryState';
+
+export const useCategoryDeleteList = () => useRecoilValue(categoryIdsArray);
+export const useResetCategoryDeleteList = () => useResetRecoilState(categoryIdsArray);
