+++ src/application/store/category/categoryState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const categoryIdsArray = atom<Array<number>>({
+  key: 'categoryIdsArray',
+  default: [],
+});
