+++ src/application/store/scrap/scrapHook.ts
@@ -1,0 +1,6 @@
+import { useRecoilValue, useResetRecoilState } from 'recoil';
+
+import { scrapIdsArray } from './scrapState';
+
+export const useScrapDeleteList = () => useRecoilValue(scrapIdsArray);
+export const useResetScrapDeleteList = () => useResetRecoilState(scrapIdsArray);
